AAE-12447 extract imageutils to common utils
