+++ activiti-cloud-acceptance-tests/activiti-cloud-acceptance-tests-core/pom.xml
@@ -43,14 +43,8 @@
       <artifactId>awaitility</artifactId>
     </dependency>
     <dependency>
-      <groupId>org.apache.xmlgraphics</groupId>
-      <artifactId>batik-transcoder</artifactId>
-      <version>${batik.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.xmlgraphics</groupId>
-      <artifactId>batik-codec</artifactId>
-      <version>${batik.version}</version>
+      <groupId>org.activiti.cloud</groupId>
+      <artifactId>activiti-cloud-services-common-util</artifactId>
     </dependency>
   </dependencies>
 </project>
