+++ activiti-cloud-service-common/pom.xml
@@ -48,6 +48,7 @@
     <jsonwebtoken.version>0.9.1</jsonwebtoken.version>
     <!-- override resteasy vulnerable 0.8.3 version of org.apache.james dependencies -->
     <james.apache-mime4j.version>0.8.9</james.apache-mime4j.version>
+    <batik.version>1.14</batik.version>
   </properties>
   <dependencyManagement>
     <dependencies>
